Split array and dict helpers into separate files
